+++ packages/core/src/connectivity/scp-cf/destination/destination-accessor-loading-precendence.spec.ts
@@ -76,7 +76,7 @@ describe('destination loading precedence', () => {
         { cacheVerificationKeys: false }
       )
     ).rejects.toThrowErrorMatchingInlineSnapshot(
-      '"Unable to get access token for \\"destination\\" service. No service instance of type \\"destination\\" found."'
+      '"No binding to an XSUAA service instance found. Please make sure to bind an instance of the XSUAA service to your application."'
     );
   });
 });