+++ CHANGELOG.md
@@ -14,6 +14,50 @@
 
 ## Compatibility Notes
 
+- [core] Tighten the return type of the `getAuthHeaders()` and `buildAuthorizationHeaders()` functions.
+
+## New Functionality
+
+-
+
+## Improvements
+
+- [openapi-generator] Replace `oclif` library with `yargs`. This is a refactoring without functional changes.
+
+## Fixed Issues
+
+- [odata] Fix parsing of `Edm.Time` missing some parts of the time e.g. `PT1H21S`.
+- [util] `setGlobalLogLevel(level)` sets log levels of all existing loggers globally.
+- [util] Fix `equal(item1, item2)` method to return `false` when either object is `null` or `undefined`.
+
+# 1.50.0
+
+Release Date: TBD<br>
+API Docs: https://sap.github.io/cloud-sdk/api/1.50.0<br>
+Blog: TBD<br>
+
+## New Functionality
+
+- [connectivity] Allow token forwarding for local destination stored in the environment variables.
+
+## Improvements
+
+- [generator] Include and exclude options added to the `transpileDirectory()` method
+
+## Fixed Issues
+
+- [core] Fix a type error, when `moment()` is used in filtering an OData field of type `Edm.DateTimeOffset`
+- [core] Fix missing `Proxy-Authorization` header if custom authorization headers are set.
+- [core] Remove wrongly given warning that custom headers are given and will overwrite destination headers.
+
+# 1.49.0
+
+Release Date: TBD<br>
+API Docs: https://sap.github.io/cloud-sdk/api/1.49.0<br>
+Blog: TBD<br>
+
+## Compatibility Notes
+
 - [core] Stop using client credentials to fetch public keys for token verification.
 - [core] Replace own implementation for requesting user tokens and client credential tokens from the XSUAA service with the `@sap/xssec` library.
 
@@ -23,12 +67,14 @@
 
 ## Improvements
 
--
+- [core] Reduce the number of warnings when selecting an XSUAA instance.
 
 ## Fixed Issues
 
 - [core] Fix verification key retrieval to match the given key id (`kid`).
 - [core] Fix the missing export for files under `http-agent` directory.
+- [core] Support using custom http(s)Agent when sending CSRF token requests.
+- [core] Fix URL encoding for `$orderby` query parameter of OData requests.
 
 # 1.48.1